feat: implementação de registro de vendas
Adicionado o serviço `VendaService` e o controlador `VendaController`. Atualizadas as views `Cadastro.cshtml`, `Index.cshtml` e `Editar.cshtml` para gerenciar vendas, incluindo formulários e cálculos de totais.

+++ Entities/VendaProdutos.cs
@@ -7,7 +7,7 @@ namespace SistemaVenda.Entities
     {
         public int CodigoVenda { get; set; }
         public int CodigoProduto { get; set; }
-        public double Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
         public decimal ValorUnitario { get; set; }
         public decimal ValorTotal { get; set; }
         public Produto Produto { get; set; }
